Added cross platform audio example
Users can upload audio clips from their device using BridgeIt. The 10
most recent audio clips will be shared on the page for everyone to
listen to. The device / OS version will be displayed on the page beside
the audio clip so they can see which device the clip was created on.